+++ src/Resources/Guests/Baskets.ts
@@ -1,13 +1,13 @@
-import AbstractResource from '../../AbstractResource';
+import AbstractResource from '../AbstractResource';
 
 /**
  * _Used for **guest** baskets only._
- * - Getting a guest basket: `$whirli.guest.basket.get(basketId);`
- * - Adding a basketLine: `$whirli.guest.basket.addLine({ basketId, productVariantId, quantity });`
- * - Updating a basketLine: `$whirli.guest.basket.updateLine({ basketLineId, quantity });`
- * - Deleting a basketLine `$whirli.guest.basket.removeLine(basketLineId);`
+ * - Getting a guest basket: `client.guests.baskets.get(basketId);`
+ * - Adding a basketLine: `client.guests.baskets.addLine({ basketId, productVariantId, quantity });`
+ * - Updating a basketLine: `client.guests.baskets.updateLine({ basketLineId, quantity });`
+ * - Deleting a basketLine `client.guests.baskets.removeLine(basketLineId);`
  */
-export default class Basket extends AbstractResource {
+export default class Baskets extends AbstractResource {
     initialise(): void {
         this.resourcePath = '/basket-lines';
         this.defaultAccess = 'guest';