+++ src/Resources/Users/Toyboxs.ts
@@ -3,8 +3,8 @@ import AbstractResource from '../AbstractResource';
 /**
  * _Updating a user's toybox_
  *
- * - `$whirli.user.toybox.designateToyForReturn(stockId);`
- * - `$whirli.user.toybox.undesignateToyForReturn(stockId);`
+ * - `client.user.toybox.designateToyForReturn(stockId);`
+ * - `client.user.toybox.undesignateToyForReturn(stockId);`
  */
 export default class Toybox extends AbstractResource {
     initialise(): void {